Have NetworkBaseCache
